+++ UnityProject/Packages/YooAsset/Runtime/FileSystem/DefaultEditorFileSystem/DefaultEditorFileSystem.cs
@@ -66,12 +66,12 @@ namespace YooAsset
             var operation = new DEFSRequestPackageVersionOperation(this);
             return operation;
         }
-        public virtual FSClearCacheFilesOperation ClearCacheFilesAsync(PackageManifest manifest, string clearMode, object clearParam)
+        public virtual FSClearCacheFilesOperation ClearCacheFilesAsync(PackageManifest manifest, ClearCacheFilesOptions options)
         {
             var operation = new FSClearCacheFilesCompleteOperation();
             return operation;
         }
-        public virtual FSDownloadFileOperation DownloadFileAsync(PackageBundle bundle, DownloadParam param)
+        public virtual FSDownloadFileOperation DownloadFileAsync(PackageBundle bundle, DownloadFileOptions options)
         {
             throw new System.NotImplementedException();
         }
@@ -94,11 +94,11 @@ namespace YooAsset
         {
             if (name == FileSystemParametersDefine.ASYNC_SIMULATE_MIN_FRAME)
             {
-                _asyncSimulateMinFrame = (int)value;
+                _asyncSimulateMinFrame = Convert.ToInt32(value);
             }
             else if (name == FileSystemParametersDefine.ASYNC_SIMULATE_MAX_FRAME)
             {
-                _asyncSimulateMaxFrame = (int)value;
+                _asyncSimulateMaxFrame = Convert.ToInt32(value);
             }
             else
             {