+++ UnityProject/Packages/YooAsset/Runtime/InitializeParameters.cs
@@ -38,6 +38,10 @@ namespace YooAsset
     /// </summary>
     public abstract class InitializeParameters
     {
+        /// <summary>
+        /// 同时加载Bundle文件的最大并发数
+        /// </summary>
+        public int BundleLoadingMaxConcurrency = int.MaxValue;
     }
 
     /// <summary>
@@ -83,6 +87,6 @@ namespace YooAsset
         /// 文件系统初始化参数列表
         /// 注意：列表最后一个元素作为主文件系统！
         /// </summary>
-        public List<FileSystemParameters> FileSystemParameterList;
+        public readonly List<FileSystemParameters> FileSystemParameterList = new List<FileSystemParameters>();
     }
 }