+++ UnityProject/Packages/YooAsset/Runtime/ResourcePackage/PackageManifest.cs
@@ -138,9 +138,6 @@ namespace YooAsset
             if (string.IsNullOrEmpty(location))
                 return string.Empty;
 
-            if (LocationToLower)
-                location = location.ToLower();
-
             if (AssetPathMapping1.TryGetValue(location, out string assetPath))
                 return assetPath;
             else
@@ -174,13 +171,28 @@ namespace YooAsset
         }
 
         /// <summary>
-        /// 获取资源依赖列表
+        /// 获取依赖列表
         /// 注意：传入的资源对象一定合法有效！
         /// </summary>
         public PackageBundle[] GetAllDependencies(PackageAsset packageAsset)
         {
             List<PackageBundle> result = new List<PackageBundle>(packageAsset.DependBundleIDs.Length);
             foreach (var dependID in packageAsset.DependBundleIDs)
+            {
+                var dependBundle = GetMainPackageBundle(dependID);
+                result.Add(dependBundle);
+            }
+            return result.ToArray();
+        }
+
+        /// <summary>
+        /// 获取依赖列表
+        /// 注意：传入的资源包对象一定合法有效！
+        /// </summary>
+        public PackageBundle[] GetAllDependencies(PackageBundle packageBundle)
+        {
+            List<PackageBundle> result = new List<PackageBundle>(packageBundle.DependBundleIDs.Length);
+            foreach (var dependID in packageBundle.DependBundleIDs)
             {
                 var dependBundle = GetMainPackageBundle(dependID);
                 result.Add(dependBundle);
@@ -291,9 +303,6 @@ namespace YooAsset
                 YooLogger.Error("Failed to mapping location to asset path, The location is null or empty.");
                 return string.Empty;
             }
-
-            if (LocationToLower)
-                location = location.ToLower();
 
             if (AssetPathMapping1.TryGetValue(location, out string assetPath))
             {