+++ UnityProject/Packages/YooAsset/Runtime/ResourcePackage/PlayMode/PlayModeImpl.cs
@@ -97,9 +97,9 @@ namespace YooAsset
         /// <summary>
         /// 清理缓存文件
         /// </summary>
-        ClearCacheFilesOperation IPlayMode.ClearCacheFilesAsync(string clearMode, object clearParam)
+        ClearCacheFilesOperation IPlayMode.ClearCacheFilesAsync(ClearCacheFilesOptions options)
         {
-            var operation = new ClearCacheFilesOperation(this, clearMode, clearParam);
+            var operation = new ClearCacheFilesOperation(this, options);
             return operation;
         }
 
@@ -176,7 +176,17 @@ namespace YooAsset
                 throw new Exception("Should never get here !");
 
             // 注意：如果清单里未找到资源包会抛出异常！
-            var depends = ActiveManifest.GetAllDependencies(assetInfo.Asset);
+            PackageBundle[] depends;
+            if (assetInfo.LoadMethod == AssetInfo.ELoadMethod.LoadAllAssets)
+            {
+                var mainBundle = ActiveManifest.GetMainPackageBundle(assetInfo.Asset);
+                depends = ActiveManifest.GetAllDependencies(mainBundle);
+            }
+            else
+            {
+                depends = ActiveManifest.GetAllDependencies(assetInfo.Asset);
+            }
+
             List<BundleInfo> result = new List<BundleInfo>(depends.Length);
             foreach (var packageBundle in depends)
             {